+ PatientDAO + H2Database + JPA

+++ src/main/java/com/example/demo/service/PatientService.java
@@ -5,6 +5,7 @@ import com.example.demo.model.PasswordRequest;
 import com.example.demo.model.Patient;
 import com.example.demo.model.PatientDTO;
 import com.example.demo.model.PatientMapper;
+import com.example.demo.repository.PatientDAO;
 import com.example.demo.repository.PatientRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,14 +17,18 @@ import java.util.List;
 public class PatientService {
     private final PatientRepository patientRepository;
     private final PatientMapper patientMapper;
+    private final PatientDAO patientDAO;
 
     public List<PatientDTO> getPatients() {
         return patientMapper.map(patientRepository.getPatients());
     }
 
     public PatientDTO getPatientByEmail(String email) {
-        return patientMapper.map(patientRepository.getPatientByEmail(email)
-                .orElseThrow(() -> new PatientNotFoundException("Patient with email: " + email + " does not exist")));
+        PatientDTO patientDTO = patientMapper.map(patientDAO.findByEmail(email));
+        if (patientDTO == null) {
+            throw new PatientNotFoundException("Patient with email: " + email + " does not exist");
+        }
+        return patientDTO;
     }
 
     public void addPatient(Patient patient) {
@@ -31,20 +36,26 @@ public class PatientService {
     }
 
     public void deletePatientByEmail(String email) {
-        Patient patient = patientRepository.getPatientByEmail(email)
-                .orElseThrow(() -> new PatientNotFoundException("Patient with email: " + email + " does not exist"));
+        Patient patient = patientDAO.findByEmail(email);
+        if (patient == null) {
+            throw new PatientNotFoundException("Patient with email: " + email + " does not exist");
+        }
         patientRepository.deletePatient(patient);
     }
 
     public void editPatient(String email, Patient newPatientData) {
-        Patient patient = patientRepository.getPatientByEmail(email)
-                .orElseThrow(() -> new PatientNotFoundException("Patient with email: " + email + " does not exist"));
+        Patient patient = patientDAO.findByEmail(email);
+        if (patient == null) {
+            throw new PatientNotFoundException("Patient with email: " + email + " does not exist");
+        }
         patientRepository.editPatient(patient, newPatientData);
     }
 
     public void editPassword(String email, PasswordRequest newPassword) {
-        Patient patient = patientRepository.getPatientByEmail(email)
-                .orElseThrow(() -> new PatientNotFoundException("Patient with email: " + email + " does not exist"));
+        Patient patient = patientDAO.findByEmail(email);
+        if (patient == null) {
+            throw new PatientNotFoundException("Patient with email: " + email + " does not exist");
+        }
         patientRepository.editPassword(patient, newPassword.newPassword());
     }
 }
